fix: #133 支持Github系列-修复发布路径问题
